extract html logic
